Add scatter-gather-single-cell utility
This patch adds the method to query a single cell using
the scatter-gather utility. Although we do have the
target_cell function to target a specific cell, in situations
where we need to handle a down cell (like for nova show), it
is advantageous to have this wrapper which in turn only calls
the scatter-gather-cells utility.

Change-Id: I31f2ab499aed795812a99b116cec3663760297e3