Clarify which context is used by do_query()
In get_records_sorted() we define a closure for convenience, which
takes a targeted cell context. That is actually named the same as
the context provided to get_records_sorted() itself and eclipses it.
That is likely to confuse someone at some point, this makes it
use the "child context" or "cctx" convention for clarity.

This is just a refactor - no test change means it works.

Change-Id: I9b48400fa1daeb641553681019a5bb894e98a2c9