Correct the release notes related to nova-consoleauth
The release notes said it was okay not to run the nova-consoleauth
service in Rocky, but that's not true because the Rocky code is storing
new console authorization tokens in both the database backend and the
existing nova-consoleauth backend. The use of nova-consoleauth will be
discontinued in Stein (for non-cells v1). We can't remove
nova-consoleauth until we remove cells v1.

Closes-Bug: #1788470

Change-Id: Ibbdc7c50c312da2acc59dfe64de95a519f87f123
(cherry picked from commit 4f01f4ff88de571218a36ba7c4e998296a7b52a4)