Remove noisy DEBUG log
There is a DEBUG log coming from the notification sending code path
that tries to populate the notification payload from a nova OVO and
encounters an unset / non-lazy loadable field in the source OVO.
As there are fields like MonitorMetric.numa_membw_values that are
notorious unset in many circumstances but used in the MetricPayload
we got excessive DEBUG logging from this code path.

This patch removes the offending logging statement as soo far we failed
to found cases based on this log where anything else we could do than
what the code anyhow does, defaulting the payload field to None.

Change-Id: I680e1d9860d839a76e229c4891933872388eb050
Closes-Bug: #1788634