Other host allocs may appear in gafpt during evac
The report client's get_allocations_for_provider_tree tries pretty hard
to retrieve only allocations relevant to the compute on which it's
running. But it was pointed out during review [1][2] that it's possible
to get another host's allocations if the instance is in the middle of an
evacuate. So call this out in a couple of places.

[1] https://review.openstack.org/#/c/584648/20/nova/scheduler/client/report.py@2002
[2] https://review.openstack.org/#/c/584648/21/nova/scheduler/client/report.py@2063

Change-Id: I94ecdcaf8692f937082ad740218bea8d08c1a87d