Log the operation when updating generation in ProviderTree
Today when a provider generation is updated we get a mostly
unhelpful message in the logs like:

  Updating resource provider bc0f4c8e-96e2-4e68-a06e-f7e0ac9aac6b
  generation from 0 to 1

What we really want to know with that is in what context did the
generation change, i.e. did inventory or traits change?

This adds the actual operation to the log message when generation
changes.

Change-Id: I9b61f1dfb8db06e02ff79e19c055c673094a4ed2
Related-Bug: #1789654