Add debug logs for when provider inventory changes
Because of the ProviderTree caching involved in both
the ResourceTracker and SchedulerReportClient, it can
be hard to know if inventory changes are being correctly
reported to the placement service as expected, especially
with things like configurable allocation ratios. This
adds debug logs to the SchedulerReportClient and
ProviderTree to determine if inventory has changed and
when it's flushed back to placement.

Change-Id: Ia6ab3ab4dea08b479bb6b794f408fd2e6f678c50
Related-Bug: #1789654