libvirt: implement reshaper for vgpu
This implements the reshaper routine for the libvirt driver
to detect and move, if necessary, VGPU inventory and allocations
from the root compute node provider to a child provider of
VGPU resources. The reshape will be performed on first start
of nova-compute with this code.

For a fresh compute node deploy, no reshaping will be necessary
and the VGPU inventory will start on the child provider.

Part of blueprint reshape-provider-tree
Part of blueprint vgpu-stein

Co-Authored-By: Sylvain Bauza <sbauza@free.fr>

Change-Id: I511d26dc6487fadfcf22ba747abd385068e975a4