Document unset/reset wrinkle for *_allocation_ratio options
This is a follow up to I43a23a3290db0c835fed01b8d6a38962dc61adce
which makes the cpu/disk/ram_allocation_ratio config "sticky" in
that once set to a non-default value, it is not possible to reset
back to the default behavior (when config is 0.0) on an existing
compute node record by unsetting the option from nova.conf. To
reset back to the defaults, the non-0.0 default would have to be
explicitly put into config, so cpu_allocation_ratio=16.0 for example.

Alternatively operators could delete the nova-compute service
record via the DELETE /os-services/{service_id} REST API and
restart the nova-compute service to get a new compute_nodes record,
but that workaround is messy and left undocumented in config.

Change-Id: I908615d82ead0f70f8e6d2d78d5dcaed8431084d
Related-Bug: #1789654