Update docs for live_migration_progress_timeout option
Change I1a1143ddf8da5fb9706cf53dbfd6cbe84e606ae1 in Ocata
deprecated the libvirt.live_migration_progress_timeout
and disabled it by default. This change updates the config
option help to refer to the bug so people don't have to hunt
for it via git history, and also touches up the admin docs.
In the one doc, mention of the option is removed altogether
because it basically says, "here is a loaded gun, but don't
use it!". It's better to just not mention the option at all.

Change-Id: I33f3d508a2af6c94435f86ac740cf24b97dba76e
Related-Bug: #1644248