Fix docs and add functional test for AggregateMultiTenancyIsolation
The docs for AggregateMultiTenancyIsolation were misleading in that
tenants are not restricted to hosts only in a tenant-isolated
aggregate. It's the opposite: hosts in the tenant-isolated aggregate
are only available for tenants configured for that aggregate.

This fixes the docs including an example for clarification, and also
adds a functional test to show the behavior of the filter.

Change-Id: Ic55b88e7ad21ab5b7ad063eac743ff9406aae559
Related-Bug: #1771523