ironic: stop hammering ironic API in power sync loop
Use our node cache to look up the node for an instance, if we have it.
If we can't find it, fall back to hitting ironic's API.

This should be relatively up-to-date info, as it is refreshed on every
resource tracker loop.

Closes-Bug: #1793556
Change-Id: I0069cbc327d952d42dbb8fe54949faab89995a7e
(cherry picked from commit 9d5fb1b58e908ccacbbbf29341918d0b0588a36f)