Fix mock.patch usage in unit tests
There are various places where a patcher was started but never stopped
or the test stopped the wrong thing. This causes that the mocking is not
removed at the end of the test case execution. So the subsequnet test
cases executed by the same executor will see the same mocked function.
Depending on the test case exection order it can lead to intermittent
test failures.

Change-Id: I8e66154c19c125f3093c8a1990b0c79332996560