Mention SR-IOV cold migration limitation in admin docs
It wasn't possible to perform a cold migration with
SR-IOV devices attached to a server until Newton.

This adds a mention of that limitation in the admin
docs on PCI pass-through. Since the note this is going
into already has a known limitation, it's re-formatted
to be a list for readability and emphasis.

Change-Id: I983107b5bf43cd46adfc8ca02e20dd30cb0b404c
Related-Bug: #1512880