Ignore VirtDriverNotReady in _sync_power_states periodic task
Change Ib0ec1012b74e9a9e74c8879f3feed5f9332b711f introduced
a new VirtDriverNotReady exception which the ironic driver raises
when asked to retrieve a list of nodes and ironic-api is not
available, like if nova-compute is started before ironic-api.
This is normal and meant to be self-healing, but we can get it
in other periodic tasks besides update_available_resource which
leads to ugly exception traces on startup in the logs. This adds
handling for the exception in the _sync_power_states periodic
task.

Change-Id: Iaf29b9e7a92705ac8a2e7ef338b92f7f1203506d
Closes-Bug: #1793768
(cherry picked from commit 6eb32bc40340fed631b9fce1245326e0ebc1c540)