Revert "Make host_aggregate_map dictionary case-insensitive"
This reverts commit 0dc0db932e3ad5ad911f2072015cb9854f6e4e23.

The original change caused our host state processing to be inconsistent
with our own hypervisors API. Automation tooling that used our API
to add hosts to aggregates would fail silently. We are reverting
this and will propose a check on the aggregate host add action
which will confirm the case-sensitive mapping of the host being
added, which is what we should have done in the first place.

Change-Id: Ibd44ba9de5680958f55f0ae6325cfc33dabadc4c
Closes-Bug: #1793747