Enforce case-sensitive hostnames in aggregate host add
If we are using a case-insensitive backend database like mysql, a
user can request an aggregate host add with a non-matching hostname
and we will not signal failure. We will create a mapping which will
not actually include that host in the aggregate, which will be
confusing later. This change makes us fail if the host name does not
match exactly, which is the same behavior as if we are using a
case-sensitive backend (like postgres).

Change-Id: I597dee74d33de337913eddda74ab056fbf81a23c
Closes-Bug: #1709260
(cherry picked from commit c8e2de668434861b87495e0a0f715b2f90d8ec05)