Handle missing marker during online data migration
During upgrade the instance used by the request spec marker could be
deleted and purged between sessions. This would cause the database
online data migration to fail as the marker instance couldn't be found.

Fix by handling the MarkerNotFound exception and re-trying without the
marker. This will go through all the instances and reset the marker when
done.

Closes-Bug: #1793419
Change-Id: If96e3d038346f16cc93209bccf3db028bacfe59b
Signed-off-by: Jack Ding <jack.ding@windriver.com>