api-ref: add 'migrations' param to GET /os-migrations
The response parameter table for the GET /os-migrations
API reference was missing the top level "migrations"
parameter which is the list of migration objects.

While in here, fix the fake "Done" status to "done"
in the sample. There is nothing in nova that uses a
status of "Done". The evacuate operation uses "done".

Change-Id: I1e193237e9aec05395cd2760d96e70db7791409b