Revert "Make host_aggregate_map dictionary case-insensitive"
This reverts commit 8f5fbd83de3759407b8f928159067211af156d60.

The original change caused our host state processing to be inconsistent
with our own hypervisors API. Automation tooling that used our API
to add hosts to aggregates would fail silently. We are reverting
this and will propose a check on the aggregate host add action
which will confirm the case-sensitive mapping of the host being
added, which is what we should have done in the first place.

NOTE(mriedem): Since this is a revert of a backport to pike,
we don't cherry pick the revert from queens, but this is using
the same change ID as the revert in queens.

Change-Id: Ibd44ba9de5680958f55f0ae6325cfc33dabadc4c
Closes-Bug: #1793747