remove commented-out code
These comments seem to have been there for quite some time. What they are
telling is simply at one time these columns were constrained by a
hosts.id foreign key. But this is not the case any longer. So keeping these
comments there is not so helpful and somewhat misleading.

Change-Id: I2bd71ce7aba839f5150669724875c237d2a86755