Fix stacktraces with redis caching backend
If you use redis as a caching backend and you delete a server with
no consoleauth tokens you'll get a stacktrace as an empty list is
passed down to the redis client and ultimately the redis server
which responds with an error, complaining about a wrong number
of arguments for the del command.
The code now checks if the list of tokens is empty and only calls
the caching backend if there are tokens available to delete.
This also may improve performance, as it no longer hands down an
empty list.

Closes-Bug: #1794812
Change-Id: Iffdd4e251bfa2bac1bfd49498e32b738843709de
(cherry picked from commit e5884a8fd51b735cf870f818d2a828163006e18a)