Use INFO for logging no allocation candidates
If the placement service returns no allocation candidates, the
log message is done in DEBUG leaving no information for the
operator to have visiblity on why the NoValidHost was raised.

If we do get something from the placement service that results
in 0 compute nodes after going through filters, we do log a
message in INFO.  This is to add consistency for a message
in INFO if we have no allocation candidates.

Closes-Bug: #1794811
Change-Id: Ief2277b3a973dd2f947f354f844ce1c6c4d6026a