Fix neutron-tempest-linuxbridge irrelevant-files
Change Ibce77d3442e21bbd5f5ce379c203542f1f31ce9e regressed
the irrelevant-files section of the neutron-tempest-linuxbridge
job which makes it run on all code changes to nova. When we
added the job to nova's check queue runs, we specifically
wanted to minimize its use to only changes in the nova/network
and libvirt/vif modules. This fixes the regression by putting
back the old irrelevant-files for that job and leaving a note
for future reviewers.

Change-Id: I00ec0282fce1a05fd314fd34bf429416a3cc9ace