Handle IndexError in _populate_neutron_binding_profile
This fixes the code that was blindly pop'ing an entry
from an empty list of PCI devices claimed by the instance.
It's not exactly clear how we can get into this situation,
presumably there was a failure in the actual PCI device
claim logic in the ResourceTracker - maybe related to the
configured PCI passthrough whitelist. Regardless, we should
handle the empty PCI device list in this method and raise
an appropriate exception to fail the build on this host.

Change-Id: I401bb74cf6e17c2b72cc62bf8ec03ec58238c44a
Closes-Bug: #1795064
(cherry picked from commit 035708c37d587e4c5ede7fe80270bdbff98016ac)