api-ref: Move the evacuate action to admin action
The evacuate action can be performed by administrators only by default.
But there is the description of the evacuate action in
the "Servers - run an action (servers, action)" section.

It should be in the "Servers - run an administrative action
(servers, action)" section.
So move to the description of the evacuate action
into the administrative action section.

Change-Id: Iafdd77c14c702af24dbb3f5a3adecc45b3014c48
Closes-Bug: #1796054