api-ref: Remove a description in servers-actions.inc
In the "Servers - run an action (servers, action)" section,
it says that you can attach a volume to a server.
But there are no APIs to attach a volume to a server
in the section.
The API is in the "Servers with volume attachments
(servers, os-volume_attachments)" section.

So remove the description in
the "Servers - run an action (servers, action)" section.

TrivialFix
Change-Id: I0bbfa98385930e21ac12815675851036c9a77621